WIP - refactor
